% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{corOrange}
\alias{corOrange}
\title{Correlations in Orange tree data}
\usage{
corOrange(orangeDS)
}
\value{
Correlation calculation for each tree.
}
\description{
Given orange tree data "Orange", the function returns the correlation between a Tree's age and circumference for each tree in the dataset.
Packages used: roxygen2, devtools, ggplot2, dplyr
}
\examples{
corOrange(Orange)
}
