% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{plotOrange}
\alias{plotOrange}
\title{Plot of "Orange" tree data.}
\usage{
plotOrange(orangeDS)
}
\value{
Scatter Plot of Tree data with line of best fit.
}
\description{
Given orange tree data "Orange", the function creates a scatter plot of data with line of best fit for each tree (age on x-axis, circumference on y-axis).
Packages used: roxygen2, devtools, ggplot2, dplyr
}
\examples{
plotOrange(Orange)
}
